+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- License settings showing stale app version instead of current version
+- Deactivation not reporting server-side failure to user
+
 ## [0.31.1] - 2026-04-12
 
 ### Fixed